reimplemented java/c++ communication to fix random jvm crashes berkelium has thread access checks now included berkelium-java.h in all cpp files for eclipse syntax highlighting support

+++ src/main/native/SingleThreadBerkelium.cpp
@@ -1,14 +1,9 @@
-#include <org_berkelium_java_impl_SingleThreadBerkelium.h>
-
-#if defined(WIN32) || defined(WIN64)
-#include <windows.h>
-#else
-#include <stdlib.h>
-#endif
+#include "berkelium-java.h"
 
 JNIEXPORT void JNICALL Java_org_berkelium_java_impl_SingleThreadBerkelium__1init(JNIEnv* env, jclass, jstring path, jstring berkeliumPath)
 {
-	Berkelium_Java_Env jEnv(env);
+	Java* java = Java::getOrCreateJava(env);
+	if(java == NULL) return;
 	jboolean iscopy;
 	const char* cPath = (const char*)env->GetStringUTFChars(path, &iscopy);
 #if defined(WIN32) || defined(WIN64)
@@ -29,24 +24,25 @@ JNIEXPORT void JNICALL Java_org_berkelium_java_impl_SingleThreadBerkelium__1init
 	env->ReleaseStringUTFChars(path, cPath);
 #if defined(WIN32) || defined(WIN64)
 	// doesn't compile with second argument!
-	Berkelium::init(Berkelium::FileString::empty(), jstring2WideString(berkeliumPath));
+	Berkelium::init(Berkelium::FileString::empty(), java->jstring2WideString(berkeliumPath));
 	//Berkelium::init(Berkelium::FileString::empty());
 #else
 	// doesn't compile with second argument!
-	Berkelium::init(Berkelium::FileString::empty(), jstring2String(berkeliumPath));
+	Berkelium::init(Berkelium::FileString::empty(), java->jstring2String(berkeliumPath));
 	//Berkelium::init(Berkelium::FileString::empty());
 #endif
 }
 
-//TODO private _destroy with thread access check
 JNIEXPORT void JNICALL Java_org_berkelium_java_impl_SingleThreadBerkelium_destroy(JNIEnv* env, jobject)
 {
-	Berkelium_Java_Env jEnv(env);
+	Java* java = Java::getOrCreateJava(env);
+	if(java == NULL) return;
 	Berkelium::destroy();
 }
 
 JNIEXPORT void JNICALL Java_org_berkelium_java_impl_SingleThreadBerkelium__1update(JNIEnv* env, jobject)
 {
-	Berkelium_Java_Env jEnv(env);
+	Java* java = Java::getOrCreateJava(env);
+	if(java == NULL) return;
 	Berkelium::update();
 }
